Разбивает хелперы из util.js по функциональности
