--- src/util.js
@@ -1,76 +1,0 @@
-const render = (container, component, place = `beforeend`) => {
-  switch (place) {
-    case `afterbegin`:
-      container.prepend(component.getElement());
-      break;
-    case `beforeend`:
-      container.append(component.getElement());
-      break;
-  }
-};
-
-const getRandomNumber = (min, max, round = 1) => {
-  const rand = min - 0.5 + Math.random() * (max - min + 1);
-  return Math.round((rand) * round) / round;
-};
-
-const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
-
-const getRandomArrayElement = (array) => {
-  return array[Math.floor(Math.random() * array.length)];
-};
-
-const shuffleArray = (array) => {
-  let currentIndex = array.length;
-  let temporaryValue;
-  let randomIndex;
-
-  while (currentIndex !== 0) {
-    randomIndex = Math.floor(Math.random() * currentIndex);
-    currentIndex -= 1;
-    temporaryValue = array[currentIndex];
-    array[currentIndex] = array[randomIndex];
-    array[randomIndex] = temporaryValue;
-  }
-
-  return array;
-};
-
-const getMonthName = (dateObject) => {
-  const monthNames = [
-    `January`, `February`, `March`, `April`, `May`, `June`,
-    `July`, `August`, `September`, `October`, `November`, `December`,
-  ];
-  return monthNames[dateObject.getMonth()];
-};
-
-const getRandomBoolean = () => {
-  return getRandomArrayElement([true, false]);
-};
-
-const getFormatDuration = (duration) => {
-  const hours = Math.floor(duration / 60);
-  const minutes = Math.round(duration % 60);
-  return `${hours}h ${minutes}m`;
-};
-
-const getCommentFormatDate = (date) => {
-  return `${date.getFullYear()}/${date.getMonth()}/${date.getDate()} ${date.getHours()}:${date.getMinutes()}`;
-};
-
-export {
-  render,
-  getRandomNumber,
-  createElement,
-  getRandomArrayElement,
-  shuffleArray,
-  getMonthName,
-  getRandomBoolean,
-  getFormatDuration,
-  getCommentFormatDate
-};
