+++ src/utils/render.js
@@ -1,0 +1,19 @@
+const createElement = (template) => {
+  const newElement = document.createElement(`div`);
+  newElement.innerHTML = template;
+
+  return newElement.firstChild;
+};
+
+const render = (container, component, place = `beforeend`) => {
+  switch (place) {
+    case `afterbegin`:
+      container.prepend(component.getElement());
+      break;
+    case `beforeend`:
+      container.append(component.getElement());
+      break;
+  }
+};
+
+export {createElement, render};
